upd8

+++ code/procs/jobprocs.dm
@@ -382,7 +382,7 @@
 		if (ishuman(src))
 			if (src.traitHolder && !src.traitHolder.hasTrait("immigrant"))
 				src:spawnId(rank)
-			else if (src.traitHolder && src.traitHolder.hasTrait("immigrant"))
+			if (src.traitHolder && src.traitHolder.hasTrait("immigrant"))
 				//Has the immigrant trait - they're hiding in a random locker
 				var/list/obj/storage/SL = list()
 				for(var/obj/storage/S in lockers_and_crates)
@@ -396,39 +396,50 @@
 				if(SL.len > 0)
 					src.set_loc(pick(SL))
 
-			else if(src.traitHolder?.hasTrait("pilot"))							//Has the Pilot trait - they're drifting off-station in a pod. Note that environmental checks are not needed here.
+			if (src.traitHolder && src.traitHolder.hasTrait("pilot"))							//Has the Pilot trait - they're drifting off-station in a pod. Note that environmental checks are not needed here.
+				boutput(world, "<span class='notice'>Pilot protocol engaged.</span>")			//Debug message. Future me, delete this.
 				var/list/turf/SL = list()
-				#ifdef UNDERWATER_MAP											//This part of the code executes only if the map is an underwater one.
+				#ifdef UNDERWATER_MAP															//This part of the code executes only if the map is an underwater one.
+				boutput(world, "<span class='notice'>The map is underwater.</span>")			//Debug message. Future me, delete this.
 				for(var/obj/critter/gunbot/drone/S in critters)
+					boutput(world, "<span class='notice'>Drone found.</span>")					//Debug message. Future me, delete this.
 					if(S.z == 5)
+						boutput(world, "<span class='notice'>Drone location valid.</span>")		//Debug message. Future me, delete this.
 						var/turf/T = get_turf(S)
 						SL.Add(T)
 				var/turf/TF = pick(SL)
-				if (TF)															//Sanity check.
+				if (TF)																			//Sanity check.
+					boutput(world, "<span class='notice'>Sanity check passed.</span>")			//Debug message. Future me, delete this.
 					src.set_loc(TF)
 				TF = src.loc
 				for(var/obj/critter/gunbot/drone/SD in TF)
+					boutput(world, "<span class='notice'>Spawn location cleared.</span>")		//Debug message. Future me, delete this.
 					qdel(SD)
-				var/obj/machinery/vehicle/tank/minisub/V = new/obj/machinery/vehicle/tank/minisub/civilian(src.loc)
-				qdel(V.engine)
-				V.engine = new/obj/item/shipcomponent/engine/zero
+				var/obj/machinery/vehicle/tank/minisub/V = new/obj/machinery/vehicle/tank/minisub/pilot(src.loc)
 				V.finish_board_pod(src)
+				boutput(world, "<span class='notice'>Pilot relocated successfully.</span>")		//Debug message. Future me, delete this.
 
-				#else 															//This part of the code executes only if the map is a space one.
+				#else 																			//This part of the code executes only if the map is a space one.
+				boutput(world, "<span class='notice'>The map is in space.</span>")				//Debug message. Future me, delete this.
 				for(var/obj/critter/gunbot/drone/S in critters)
-					if(S.z != 1 && isrestrictedz(S.z))
+					boutput(world, "<span class='notice'>Drone found.</span>")					//Debug message. Future me, delete this.
+					if(S.z != 1 && !isrestrictedz(S.z))
+						boutput(world, "<span class='notice'>Drone location valid.</span>")		//Debug message. Future me, delete this.
 						var/turf/T = get_turf(S)
 						SL.Add(T)
 				var/turf/TF = pick(SL)
-				if (TF)															//Sanity check.
+				if (TF)																			//Sanity check.
+					boutput(world, "<span class='notice'>Sanity check passed.</span>")			//Debug message. Future me, delete this.
 					src.set_loc(TF)
 				TF = src.loc
 				for(var/obj/critter/gunbot/drone/SD in TF)
+					boutput(world, "<span class='notice'>Spawn location cleared.</span>")		//Debug message. Future me, delete this.
 					qdel(SD)
-				var/obj/machinery/vehicle/miniputt/V = new/obj/machinery/vehicle/miniputt(src.loc)
+				var/obj/machinery/vehicle/miniputt/V = new/obj/machinery/vehicle/miniputt/pilot(src.loc)
 				qdel(V.engine)
 				V.engine = new/obj/item/shipcomponent/engine/zero
 				V.finish_board_pod(src)
+				boutput(world, "<span class='notice'>Pilot relocated successfully.</span>")		//Debug message. Future me, delete this.
 				#endif
 
 			if (prob(10) && islist(random_pod_codes) && random_pod_codes.len)
